mitosis training

+++ src/napatrackmater/Trackvector.py
@@ -1198,13 +1198,13 @@ def compute_covariance_matrix(track_arrays):
 class MitosisNet(nn.Module):
     def __init__(self, input_size, num_classes_class1, num_classes_class2):
         super().__init__()
-        self.conv1 = nn.Conv1d(in_channels=1, out_channels=32, kernel_size=3).to(torch.float16)  
-        self.conv2 = nn.Conv1d(in_channels=32, out_channels=64, kernel_size=3).to(torch.float16)  
+        self.conv1 = nn.Conv1d(in_channels=1, out_channels=32, kernel_size=3) 
+        self.conv2 = nn.Conv1d(in_channels=32, out_channels=64, kernel_size=3)
         self.pool = nn.MaxPool1d(kernel_size=2)
         conv_output_size = self._calculate_conv_output_size(input_size)
-        self.fc1 = nn.Linear(conv_output_size, 128).to(torch.float16)  
-        self.fc2_class1 = nn.Linear(128, num_classes_class1).to(torch.float16)  
-        self.fc3_class2 = nn.Linear(128, num_classes_class2).to(torch.float16)  
+        self.fc1 = nn.Linear(conv_output_size, 128) 
+        self.fc2_class1 = nn.Linear(128, num_classes_class1) 
+        self.fc3_class2 = nn.Linear(128, num_classes_class2) 
  
 
     def _calculate_conv_output_size(self, input_size):
@@ -1230,7 +1230,7 @@ def train_mitosis_neural_net(features_array, labels_array_class1, labels_array_c
     device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 
     X_train, X_val, y_train_class1, y_val_class1, y_train_class2, y_val_class2 = train_test_split(
-        features_array.astype(np.float16), 
+        features_array.astype(np.float32), 
         labels_array_class1.astype(np.uint8), 
         labels_array_class2.astype(np.uint8), 
         test_size=0.1,
